requirements.txt
SpeechRecognition
pymongo
python-dotenv
flask
flask-pymongo
google-cloud-texttospeech
google-cloud-speech
tomli
pydub
ffmpeg
pymongo